[#1606] Fix integer overflow during Raft log read path
Summary:
A couple of locations in the Raft log / log_reader / log_cache read path had some unsafe conversions between int64 (from OpIds) to simple ints, causing overflow issues.

Similar issue observed and addressed in Kudu as well: https://github.com/apache/kudu/commit/8363b74506f8513e2fa9dbf772e30d0abce4e444

Test Plan: New tests and full Jenkins run.

Reviewers: amitanand, sergei, kannan, mikhail

Reviewed By: mikhail

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D6798

+++ src/yb/consensus/log_reader.h
@@ -118,7 +118,7 @@ class LogReader {
       const int64_t up_to,
       int64_t max_bytes_to_read,
       ReplicateMsgs* replicates) const;
-  static const int kNoSizeLimit;
+  static const int64_t kNoSizeLimit;
 
   // Look up the OpId for the given operation index.
   // Returns a bad Status if the log index fails to load (eg. due to an IO error).
